feat: add navigation example
