+++ example/src/components/sheets/BlankSheet.tsx
@@ -2,7 +2,7 @@ import { forwardRef, type Ref } from 'react'
 import { StyleSheet, Text } from 'react-native'
 import { TrueSheet, type TrueSheetProps } from '@lodev09/react-native-true-sheet'
 
-import { $WHITE_TEXT, DARK, SPACING } from '../utils'
+import { $WHITE_TEXT, DARK, SPACING } from '../../utils'
 
 interface BlankSheetProps extends TrueSheetProps {}
