+++ example/src/components/sheets/FlatListSheet.tsx
@@ -2,8 +2,9 @@ import { forwardRef, useRef, type Ref } from 'react'
 import { FlatList, View, type ViewStyle } from 'react-native'
 import { TrueSheet, type TrueSheetProps } from '@lodev09/react-native-true-sheet'
 
-import { DARK, DARK_GRAY, INPUT_HEIGHT, SPACING, times } from '../utils'
-import { DemoContent, Input } from '../components'
+import { DARK, DARK_GRAY, INPUT_HEIGHT, SPACING, times } from '../../utils'
+import { Input } from '../Input'
+import { DemoContent } from '../DemoContent'
 
 interface FlatListSheetProps extends TrueSheetProps {}
