+++ example/src/components/sheets/ScrollViewSheet.tsx
@@ -2,8 +2,9 @@ import { forwardRef, useRef, type Ref } from 'react'
 import { ScrollView, type ViewStyle } from 'react-native'
 import { TrueSheet, type TrueSheetProps } from '@lodev09/react-native-true-sheet'
 
-import { FOOTER_HEIGHT, SPACING, times } from '../utils'
-import { DemoContent, Footer } from '../components'
+import { FOOTER_HEIGHT, SPACING, times } from '../../utils'
+import { Footer } from '../Footer'
+import { DemoContent } from '../DemoContent'
 
 interface ScrollViewSheetProps extends TrueSheetProps {}
