+++ example/src/types.ts
@@ -1,0 +1,5 @@
+export type AppStackParamList = {
+  Map: undefined
+  Navigation: undefined
+  Child: undefined
+}
